Fix NoneType attribute error in mainframe.py
From what I can tell a callback from QT was being hit before the frame
value on the room was populated.  This caused the method to try and
invoke the scroll method on a NoneType object, which of course fails.

This fixes the issue by using the frame object of the parent in the
method, which is likely the correct object as this only occurs during
the creation of a frame representing a chat room.