Check for DBUS Notifications service presence
Instead of checking for the presence of particular environment
variables, set by Desktop Environments, we should check if there is
any service on DBus that listens to org.freedesktop.Notifications.

This is important for people who use different services that provide
notification features like Dunst[1].

See [2] to check a more detailed list.

1: https://github.com/knopwob/dunst
2: https://wiki.archlinux.org/index.php/Desktop_Notifications